fix: rename to tabs
